Refactoring for handling ProcessVariableDeletedEvent

+++ activiti-cloud-services-query/activiti-cloud-services-query-rest/src/main/java/org/activiti/cloud/services/query/events/handlers/ProcessVariableDeletedEventHandler.java
@@ -16,12 +16,18 @@
 
 package org.activiti.cloud.services.query.events.handlers;
 
+import java.util.Optional;
+
 import com.querydsl.core.types.dsl.BooleanExpression;
 import org.activiti.cloud.api.model.shared.events.CloudVariableDeletedEvent;
 import org.activiti.cloud.services.query.app.repository.EntityFinder;
+import org.activiti.cloud.services.query.app.repository.ProcessInstanceRepository;
+import org.activiti.cloud.services.query.app.repository.TaskVariableRepository;
 import org.activiti.cloud.services.query.app.repository.VariableRepository;
+import org.activiti.cloud.services.query.model.ProcessInstanceEntity;
 import org.activiti.cloud.services.query.model.ProcessVariableEntity;
 import org.activiti.cloud.services.query.model.QProcessVariableEntity;
+import org.activiti.cloud.services.query.model.TaskEntity;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -32,9 +38,13 @@ public class ProcessVariableDeletedEventHandler {
     private final VariableRepository variableRepository;
 
     private final EntityFinder entityFinder;
+    
+    private final ProcessInstanceRepository processRepository;
 
-    public ProcessVariableDeletedEventHandler(VariableRepository variableRepository,
+    public ProcessVariableDeletedEventHandler(ProcessInstanceRepository processRepository,
+                                              VariableRepository variableRepository,
                                               EntityFinder entityFinder) {
+        this.processRepository = processRepository;
         this.variableRepository = variableRepository;
         this.entityFinder = entityFinder;
     }
@@ -42,21 +52,27 @@ public class ProcessVariableDeletedEventHandler {
     public void handle(CloudVariableDeletedEvent event) {
         String variableName = event.getEntity().getName();
         String processInstanceId = event.getEntity().getProcessInstanceId();
-        BooleanExpression predicate = QProcessVariableEntity.processVariableEntity.processInstanceId.eq(processInstanceId)
-                .and(
-                        QProcessVariableEntity.processVariableEntity.name.eq(variableName)
-                );
-        ProcessVariableEntity variableEntity = entityFinder.findOne(variableRepository,
-                                                             predicate,
-                                                             "Unable to find variableEntity with name '" + variableName + "' for process instance '" + processInstanceId + "'");
+        Optional<ProcessInstanceEntity> findResult = processRepository.findById(processInstanceId);
         
-        // Persist into database
-        try {
-            variableRepository.delete(variableEntity);
-        } catch (Exception cause) {
-            LOGGER.debug("Error handling ProcessVariableDeletedEvent[" + event + "]",
-                         cause);
-        }       
+        // if a task was cancelled / completed do not handle this event
+        if(findResult.isPresent() && !findResult.get().getStatus().isFinalState()) {
+        
+            BooleanExpression predicate = QProcessVariableEntity.processVariableEntity.processInstanceId.eq(processInstanceId)
+                    .and(
+                            QProcessVariableEntity.processVariableEntity.name.eq(variableName)
+                    );
+            ProcessVariableEntity variableEntity = entityFinder.findOne(variableRepository,
+                                                                 predicate,
+                                                                 "Unable to find variableEntity with name '" + variableName + "' for process instance '" + processInstanceId + "'");
+            
+            // Persist into database
+            try {
+                variableRepository.delete(variableEntity);
+            } catch (Exception cause) {
+                LOGGER.debug("Error handling ProcessVariableDeletedEvent[" + event + "]",
+                             cause);
+            }   
+        }
         
     }
 }
